+++ public/Nitrocid/Misc/Writers/FancyWriters/FigletWhereColor.cs
@@ -129,7 +129,7 @@ namespace KS.Misc.Writers.FancyWriters
             try
             {
                 // Check if default console output equals the new console output text writer. If it does, write in color, else, suppress the colors.
-                ColorTools.SetConsoleColor(new Color((int)Color));
+                ColorTools.SetConsoleColor(new Color(Color));
 
                 // Actually write
                 WriteFigletWherePlain(Text, Left, Top, Return, FigletFont, Vars);
@@ -157,8 +157,8 @@ namespace KS.Misc.Writers.FancyWriters
             try
             {
                 // Check if default console output equals the new console output text writer. If it does, write in color, else, suppress the colors.
-                ColorTools.SetConsoleColor(new Color((int)ForegroundColor));
-                ColorTools.SetConsoleColor(new Color((int)BackgroundColor), true);
+                ColorTools.SetConsoleColor(new Color(ForegroundColor));
+                ColorTools.SetConsoleColor(new Color(BackgroundColor), true);
 
                 // Actually write
                 WriteFigletWherePlain(Text, Left, Top, Return, FigletFont, Vars);
